+++ Changelog.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## Changed
+
+- Made `SizeOf` impls for `Context`, `TotalSize` and `HumanBytes` unconditional
+  (were previously dependent on the `derive` feature)
+
 ## [0.1.3] - 2022-10-21
 
 ## Added